restructuring map
